% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_object.R
\docType{class}
\name{flamingo-class}
\alias{flamingo-class}
\title{Create an object named flamingo for data organization and an object named flamingo_prediction for result organization}
\value{
Objects named flamingo and flamingo_prediction are ready for use.
}
\description{
Create an object named flamingo for data organization and an object named flamingo_prediction for result organization
}
\keyword{flamingo_object}
