% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo_backbone.R
\name{flamingo_backbone}
\alias{flamingo_backbone}
\title{backbone reconstruction of flamingorlite
Core function for 3D genome structure reconstruction using low-rank matrix completion}
\usage{
flamingo_backbone(
  temp_folder,
  sample_rate,
  lambda,
  r,
  max_dist,
  error_threshold,
  max_iter,
  alpha,
  inf_dist
)
}
\arguments{
\item{temp_folder}{tempfolder.}

\item{sample_rate}{Fraction of available entries in Hi-C to be used during the reconstruction. Default = 0.75.}

\item{lambda}{Weights for all sampled entries. Default = 10.}

\item{r}{Weights for distance between consecutive points. Default = 1.}

\item{max_dist}{Maximum allowed distance betwee two consecutive points. Default = 0.1}

\item{error_threshold}{Error thresholds for reconstruction. Default = 1e-4.}

\item{max_iter}{Maximum iterations. Default = 500.}

\item{alpha}{Convertion factor between interaction frequency and pairwise distance. Default = -0.25.}

\item{inf_dist}{Maximun allowed distance betwee any two points. Default = 3.}
}
\value{
A flamingo_prediction object containing the fragment id and 3D coordinates
}
\description{
backbone reconstruction of flamingorlite
Core function for 3D genome structure reconstruction using low-rank matrix completion
}
\keyword{flamingo_backbone}
