% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo_basic.R
\name{flamingo_basic}
\alias{flamingo_basic}
\title{main function of flamingo
Core function for 3D genome structure reconstruction using low-rank matrix completion}
\usage{
flamingo_basic(
  input_if,
  sample_rate = 0.75,
  lambda = 10,
  r = 1,
  max_dist = 0.01,
  error_threshold = 0.001,
  max_iter = 300,
  alpha = -0.25,
  inf_dist = 2
)
}
\arguments{
\item{input_if}{Input interaction frequency matrix.}

\item{sample_rate}{Fraction of available entries in Hi-C to be used during the reconstruction. Default = 0.75.}

\item{lambda}{Weights for all sampled entries. Default = 10.}

\item{r}{Weights for distance between consecutive points. Default = 1.}

\item{max_dist}{Maximum allowed distance betwee two consecutive points. Default = 0.01}

\item{error_threshold}{Error thresholds for reconstruction. Default = 1e-3.}

\item{max_iter}{Maximum iterations. Default = 500.}

\item{alpha}{Convertion factor between interaction frequency and pairwise distance. Default = -0.25.}

\item{inf_dist}{Maximun allowed distance betwee any two points. Default = 2.}
}
\value{
A flamingo_prediction object containing the fragment id and 3D coordinates
}
\description{
main function of flamingo
Core function for 3D genome structure reconstruction using low-rank matrix completion
}
\keyword{flamingo_basic}
