% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo_main.R
\name{flamingo_main}
\alias{flamingo_main}
\title{flamingo.main_func}
\usage{
flamingo_main(
  hic_data,
  file_format,
  domain_res,
  frag_res,
  chr_name,
  normalization,
  nThread = 1,
  sample_rate = 0.75,
  lambda = 10,
  r = 1,
  max_dist = 0.01,
  alpha = -0.25,
  inf_dist = 2,
  error_threshold = 0.001,
  max_iter = 500
)
}
\arguments{
\item{hic_data}{Input Hi-C data. Now only support .hic}

\item{file_format}{Foramt of the input hic data. Now only support .hic.}

\item{domain_res}{Size of the domains in bps, e.g. 1e6. Try strawr::readHicBpResolutions() to see available resolutions.}

\item{frag_res}{Size of the fragment in bps, e.g. 5e3. Try strawr::readHicBpResolutions() to see available resolutions.}

\item{chr_name}{Name of the chromosome, e.g. chr1. Try strawr::readHicChroms() to see available chromosomes.}

\item{normalization}{Normalization method in .hic file. Try strawr::readHicNormTypes() to see available methods. Could be 'NONE'.}

\item{nThread}{Number of thread avalable for the reconstruction. Default = 1.}

\item{sample_rate}{Fraction of available entries in Hi-C to be used during the reconstruction. Default = 0.75.}

\item{lambda}{Weights for all sampled entries. Default = 10.}

\item{r}{Weights for distance between consecutive points. Default = 1.}

\item{max_dist}{Maximum allowed distance betwee two consecutive points. Default = 0.01}

\item{alpha}{Convertion factor between interaction frequency and pairwise distance. Default = -0.25.}

\item{inf_dist}{Maximun allowed distance betwee any two points. Default = 2.}

\item{error_threshold}{Error thresholds for reconstruction. Default = 1e-3.}

\item{max_iter}{Maximum iterations. Default = 500.}
}
\value{
A data.frame containing the FLAMINGO predicted 3D structure.
}
\description{
Main function of hierarchical-FLAMINGO
}
\keyword{FLAMINGO}
